+++ chaosmagpy/__init__.py
@@ -1,7 +1,8 @@
 __all__ = [
     "CHAOS", "load_CHAOS_matfile", "load_CHAOS_shcfile",
     "load_CovObs_txtfile", "load_gufm1_txtfile", "load_CALS7K_txtfile",
-    "basicConfig", "synth_values"
+    "basicConfig", "synth_values", "mjd2000", "timestamp", "dyear_to_mjd",
+    "mjd_to_dyear"
 ]
 
 __version__ = "0.11-dev"
@@ -14,5 +15,7 @@ from .chaos import (
     load_gufm1_txtfile,
     load_CALS7K_txtfile
 )
+
 from .config_utils import basicConfig
 from .model_utils import synth_values
+from .data_utils import mjd2000, timestamp, dyear_to_mjd, mjd_to_dyear